services/agent_support/pyproject.toml
[tool.poetry]
name = "hedera-rag-server"
version = "0.1.0"
description = "Lightweight Retrieval-Augmented Generation (RAG) service for the Hedera network leveraging FastMCP and LlamaIndex."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

# The actual source code lives in the nested `agent_support/hedera_rag_server` package
packages = [
    { include = "hedera_rag_server", from = "agent_support" }
]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "<1.25"
fastmcp = "^2.4.0"
fastapi = "^0.110.0"
httpx = "^0.28.1"
llama-index-core = "^0.10.20"
llama-index-embeddings-huggingface = "^0.2.0"
sentence-transformers = "^2.2.0"
# Install CPU-only build of PyTorch to keep images slim when containerising on non-macOS platforms.
torch = { version = "2.1.0+cpu", source = "pytorch-cpu", markers = "platform_system != 'Darwin'" }

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
ruff = "^0.2"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
